apiのURLを修正、mapのclickイベントをwidgetに移動
