+++ app/components/farm-field/farm-field-location.module.css
@@ -25,22 +25,22 @@
 .styled-table th, .styled-table td {
   padding: 12px 15px;
 }
+
 .styled-table thead tr {
   background-color: #009879;
   color: #ffffff;
   text-align: left;
 }
+
 .styled-table tbody tr {
   border-bottom: 1px solid #dddddd;
 }
+
 .styled-table tbody tr:nth-of-type(even) {
   background-color: #f3f3f3;
 }
-.styled-table tbody tr:last-of-type {
-  /* border-bottom: 2px solid #009879; */
-} 
+
 .styled-table tbody tr.active-row {
   font-weight: bold;
   color: #009879;
 }
-
