+++ app/routes/farm-field.tsx
@@ -1,11 +1,11 @@
 import { useState } from "react";
 import FarmFieldMap from "~/components/farm-field/farm-field-map";
 import FarmFieldLocation from "~/components/farm-field/farm-field-location";
-import styles from "~/components/farm-field/farm-field.module.css";
+import styles from "./farm-field.module.css";
 import type { Route } from "./+types/farm-field";
 
 const getFarmFieldList = async (userId: string) => {
-  const response = await fetch(`http://localhost:8000/api/farm-field/${userId}`);
+  const response = await fetch(`http://localhost:8000/api/farm_fields/${userId}`);
   if (response.status !== 200) throw response;
   const data = await response.json();
   return data;
